% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{create_element_styles}
\alias{create_element_styles}
\title{Create element styles}
\usage{
create_element_styles(opts_theme)
}
\arguments{
\item{opts_theme}{Theme options from dsvizopts}
}
\value{
Named list of strings of css styles
}
\description{
Create css styles for extra elements title, subtitle, caption and logo.
}
