% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{create_link_style}
\alias{create_link_style}
\title{Create link style}
\usage{
create_link_style(opts_theme)
}
\arguments{
\item{opts_theme}{Theme options from dsvizopts}
}
\value{
String of css style
}
\description{
Create css style of hyperlinks in table.
}
