% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{create_rctbtheme}
\alias{create_rctbtheme}
\title{Create rctbtheme}
\usage{
create_rctbtheme(opts_theme)
}
\arguments{
\item{opts_theme}{Theme options from dsvizopts}
}
\value{
reactableTheme
}
\description{
Create extra styling for fonts and colors of reactable.
}
