% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{get_font_names}
\alias{get_font_names}
\title{Get font names}
\usage{
get_font_names(opts_theme)
}
\arguments{
\item{opts_theme}{Theme options from dsvizopts}
}
\value{
Vector of strings of font names
}
\description{
Get names of google fonts in theme options. Searches for all
parameters in theme options with term \code{family}
}
