% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{get_html_logo}
\alias{get_html_logo}
\title{Get html logo without styles}
\usage{
get_html_logo(opts_theme)
}
\arguments{
\item{opts_theme}{Theme options from dsvizopts}
}
\value{
HTML object for logo
}
\description{
Get html logo without styles
}
