% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{get_hyperlink_function}
\alias{get_hyperlink_function}
\title{Get hyperlink function}
\usage{
get_hyperlink_function(opts)
}
\arguments{
\item{opts}{Options from dsvizopts}
}
\value{
Function passed to reactable cells
}
\description{
Creates function to format hyperlinks in reactable
}
