% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{get_logo_style}
\alias{get_logo_style}
\title{Get css style for html logo}
\usage{
get_logo_style(opts_theme)
}
\arguments{
\item{opts_theme}{Theme options from dsvizopts}
}
\value{
String of CSS styles for logo
}
\description{
Get css style for html logo
}
