% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctb.R
\name{rctb}
\alias{rctb}
\title{Create rctb table}
\usage{
rctb(data, ...)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{}
}
\value{

}
\description{
Wrapper function for reactable tables from the reactable package.
}
\examples{
org_name <- "public"
data <- data.frame(a = c('b', 'c'), b = c('https://google.com', 'https://google.co.uk'))

opts <- dsthemer::dsthemer_get(org_name)
opts$logo <- org_name
opts$logo_width <- 200
opts$title <- "a title"
opts$subtitle <- "a subtitle"
opts$caption <- "a caption"

do.call(rctb, list(data, opts))
}
